python/kvikio/pyproject.toml
# SPDX-FileCopyrightText: Copyright (c) 2021-2025, NVIDIA CORPORATION. All rights reserved.
# SPDX-License-Identifier: Apache-2.0

[build-system]
build-backend = "rapids_build_backend.build"
requires = [
    "rapids-build-backend>=0.4.0,<0.5.0.dev0",
    "scikit-build-core[pyproject]>=0.10.0",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.

[project]
name = "kvikio"
dynamic = ["version"]
description = "KvikIO - GPUDirect Storage"
readme = { file = "README.md", content-type = "text/markdown" }
authors = [
    { name = "NVIDIA Corporation" },
]
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
dependencies = [
    "cupy-cuda13x>=13.6.0",
    "libkvikio==25.12.*,>=0.0.0a0",
    "numpy>=1.23,<3.0a0",
    "packaging",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.optional-dependencies]
test = [
    "boto3>=1.21.21",
    "cuda-python>=13.0.1,<14.0a0",
    "moto[server]>=4.0.8",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-timeout",
    "rangehttpserver",
    "rapids-dask-dependency==25.12.*,>=0.0.0a0",
    "zarr>=3.0.0,<4.0.0; python_version >= '3.11'",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.
zarr = [
    "zarr>=3.0.0,<4.0.0; python_version >= '3.11'",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.

[project.urls]
Homepage = "https://github.com/rapidsai/kvikio"

[tool.mypy]
ignore_missing_imports = true

[tool.rapids-build-backend]
build-backend = "scikit_build_core.build"
dependencies-file = "../../dependencies.yaml"
matrix-entry = "cuda_suffixed=true;use_cuda_wheels=true"
requires = [
    "cmake>=3.30.4",
    "cython>=3.0.0",
    "libkvikio==25.12.*,>=0.0.0a0",
    "ninja",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.

[tool.scikit-build]
build-dir = "build/{wheel_tag}"
cmake.build-type = "Release"
cmake.version = "CMakeLists.txt"
minimum-version = "build-system.requires"
ninja.make-fallback = false
sdist.exclude = ["*tests*"]
sdist.reproducible = true
wheel.packages = ["kvikio"]
wheel.exclude = ["*.pyx", "CMakeLists.txt"]

[tool.scikit-build.metadata.version]
provider = "scikit_build_core.metadata.regex"
input = "kvikio/VERSION"
regex = "(?P<value>.*)"

[tool.pydistcheck]
select = [
    "distro-too-large-compressed",
]

# PyPI limit is 100 MiB, fail CI before we get too close to that
max_allowed_size_compressed = '75M'

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore:Jitify is performing a one-time only warm-up to populate the persistent cache",
    "ignore::DeprecationWarning:botocore.*",
]
markers = [
    "cufile: tests to skip if cuFile isn't available e.g. run with `pytest -m 'not cufile'`",
    "gpu: zarr-python tests requiring a GPU to run."
]
asyncio_mode = "auto"
